extract server func
