+++ src/provider/mod.rs
@@ -1,1 +1,5 @@
-pub mod rpc;
+mod rpc;
+mod server;
+
+pub use rpc::ServerData;
+pub use server::Server;
